remove meta prop prefix
